eval: add evaluation for personality
Update prompt for rephrase chat as well.

Prior to the prompt change, both variants scored 10/12 on personality. After the change, they scored 12/12.

Accuracy dropped for variant1 from 11/12 to 8/12. We also tried reducing temp to 0 in case that's causing more hallucination, but it still scored 8/12 on accuracy, so we didn't end up persisting this personality change.

However, accuracy for variant2 increased from 10/12 to 11/12, so we are keeping this change.