widgets: remove fetch-wallet, nesting of fetch commands, and notion of connected wallet
Instead, if a connected wallet is required, we throw an exception if we don't have one
and use it to return the right response to the user.

This is part of removing the nesting of fetch commands which are not that reliable
and can perform poorly due to inappropriate chaining or hallucinating.